requirements.txt
torch==2.0.1
torchvision==0.15.2
accelerate==0.23.0
diffusers==0.23.1
einops
numpy==1.24.4
Pillow==10.1.0
PyYAML
safetensors
tqdm
transformers==4.35.2
tensorboard
scipy
opencv-python
